requirements.txt
# Core Django
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0

# Database
psycopg2-binary>=2.9.0

# Environment and Configuration
python-decouple>=3.6

# CORS handling
django-cors-headers>=4.0.0

# Filtering and Search
django-filter>=23.0.0

# Image processing (for avatars)
Pillow>=10.0.0

# Redis for caching and Celery
redis>=4.5.0
celery>=5.3.0

# Development tools
django-extensions>=3.2.0

# Timezone handling
pytz>=2023.3

# Security
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
pytest-mock>=3.11.0
factory-boy>=3.3.0

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
